AIPI/server/requirements.txt
# 核心依赖
flask>=2.0.0
python-dotenv>=0.19.0
paho-mqtt>=1.6.0
langchain>=0.0.267
langchain-community>=0.0.10
langchain-openai>=0.0.2
openai>=1.6.1

# 数据存储
influxdb-client>=1.24.0

# Web服务
Flask-SocketIO>=5.3.4
Flask-Cors>=3.0.10
PyJWT>=2.7.0
Werkzeug>=2.3.6
gunicorn>=21.2.0
python-socketio>=5.8.0
eventlet>=0.33.3

# 实用工具
requests>=2.27.0
schedule>=1.2.0
uuid>=1.30
numpy>=1.20.0
pandas>=1.3.0

# 邮件和通知
python-telegram-bot>=13.15

# 测试工具
pytest>=7.4.0
pytest-cov>=4.1.0

# 其他依赖
jsonschema>=4.0.0
pymongo>=4.0.0
toml>=0.10.2
pytz>=2021.3
dataclasses-json>=0.5.7
pymongo[srv]>=4.0.0
loguru>=0.6.0
python-multipart>=0.0.5
openpyxl>=3.0.9
matplotlib>=3.5.0
python-dateutil>=2.8.2 